Fix selection of files to merge
Using `git diff left...right`, we actually get all changes from
$(git merge-base left right) to right, i.e., only one side of the merge.
The other side is retrieved with `git diff right...left`.
There might still be fast forwards, which is why we have to find
the intersection of both sets.

Also, we have to extend the diff-filter to AMR, as both sides might have
added the same file(s).

As for renames, we have to extract the old and new filenames to get
the file triple correct. This is also the reason why the return type of
get_merged_files() changed to a list of triples (l, b, r).
prepare_job() was changed accordingly.